+++ unique_sdk/CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.30] - 2025-05-21
+## [0.9.31] - 2025-05-21
 - Add function to update folder access.
 
-## [0.9.29] - 2025-05-21
-- Add function to update folder properties.
+## [0.9.30] - 2025-05-21
+- Add function to update folder ingestion config.
+
+## [0.9.29] - 2025-05-20
+- Add function to create folder paths if they do not exist.
 
 ## [0.9.28] - 2025-05-20
 - Add function to search content info. This also allows filtering content info by metadata info.